Break server/service into their own files in api
